Pizza Ordering Calculator
This application allows a person to order a pizza by determining the size of the pizza (S, M, L), if they would like to add pepperoni, and if they would like to add extra cheese. After answering the prompts it will tell them their final bill.
Mainly focused on if/elif/else statements.